refactor: 헤더 구조 개선
